requirements.txt
supabase
discord.py
python-dotenv
